feat: add static query for array of poets
